requirements.txt
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.17
django-cors-headers==3.5.0
django-rest-auth==0.9.5
django-simpleui==2020.9.26
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
PyJWT==1.7.1
PyMySQL==0.10.0
pytz==2020.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2